Avoid deauthenticating STA if the reason for freeing PMK entry isn't expiry
The PMK cache entry for a STA in the SoftAP managed by wpa_supplicant
can be freed in multiple scenarios like when a new PMK is created after
reconnection or when the the PMK Life time is expired, etc.

So avoid sending Deauth to the STA when freeing the PMK cache entry for a
specific STA in the SoftAP when attempting to replace it with a new PMK
cache entry derived as part of STA reconnection. Doing this lets the STA
reconnect to the SoftAP successfully after tearing down the connection
because of configured PMK Life time getting expired in the STA. And send
Deauth to STA only when the reason for freeing PMK cache entry is that the
configured PMK Life time got expired in the SoftAP.

The PMKSA free reasons PMK_FREE, PMK_REPLACE and PMK_EXPIRE are introduced
for SoftAP to be consistent with the PMKSA reasons defined for STA mode in
src/rsn_supp/pmksa_cache.h

(cherry picked from commit 99c91c01c585f2ed962481e137889382b9876b15)