DUT fails to connect in WPA2 when AP is configured in Transition mode.
In case AP is configured in Transition mode and DUT tries to connect to
WPA2 security, Our DUT was trying for SAE connection. Reason for same is
SAE Password is always sent to driver/FW if SAE offload is enabled.
Since both SAE password and PSK is available Firmware is trying for SAE.
Added check in supplicant to send the SAE passsword if AKM is SAE.

Fixes SWWLAN-148955

Signed-off-by: avishad verma <avishad.verma@infineon.com>